scanner-worker/requirements.txt
boto3==1.41.0
botocore==1.41.0
jmespath==1.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
regex==2025.11.3
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
